python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deeprl-selfplay-arena"
version = "0.1.0"
description = "Minimal Python training scaffold for a self-play arena RL project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pyyaml>=6.0",
  "torch>=2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# TODO: Add optional dependency groups for training, evaluation, and dev tooling.
